waterbase-backend/services/rag-ai-service/requirements.txt
# Waterbase RAG AI Service — Python Dependencies
# Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0

# AI & Embedding
google-genai==1.5.0

# Database
supabase==2.11.0

# Utilities
python-dotenv==1.0.1
requests==2.32.0
